+++ skel.c
@@ -245,7 +245,7 @@ RayPtr FindNearestRay(XButtonEvent* pevent) {
 ** Yes, I KNOW the following is a disgusting way to step through all the
 ** ray elements of the "bones" data structure. So there! Neah neah neah.
 */
-	for (stepray=&(bones.torso); stepray <= &(bones.ltricep); stepray=(RayPtr) ((int) stepray+(int) stepsize)) {
+	for (stepray=&(bones.torso); stepray <= &(bones.ltricep); stepray=(RayPtr) ((long) stepray+(long) stepsize)) {
 		newdist=(FindDistance(x, y, stepray));
 		if (newdist < olddist) {
 			retval=stepray;
